+++ src/Adapter/Amqp/Queue/AmqpQueue.php
@@ -89,7 +89,6 @@ readonly class AmqpQueue implements QueueInterface
         $envelope = $this->queue->get();
 
         if ($envelope) {
-            // @phpstan-ignore-next-line
             return new AmqpReceivedMessage($this->queue, $envelope);
         }